+++ internal/trigger/index.go
@@ -108,7 +108,7 @@ func (tr *TriggerRunnerBase) Run() {
 
 	pipelineCmd := &event.PipelineQueue{
 		Event:               event.NewExecutionEvent(),
-		PipelineExecutionID: util.NewPipelineExecutionID(),
+		PipelineExecutionID: util.NewPipelineExecutionId(),
 		Name:                pipelineName,
 		Args:                pipelineArgs,
 	}